+++ backend/src/main/java/com/example/backend/Controller/OrderController.java
@@ -2,10 +2,13 @@ package com.example.backend.Controller;
 
 import com.example.backend.DTO.OrderRequestDTO;
 import com.example.backend.DTO.OrderResponseDTO;
+import com.example.backend.Entity.Order;
 import com.example.backend.Service.MultiService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,7 +19,7 @@ import java.util.List;
 public class OrderController {
     private final MultiService multiService;
 
-    @PostMapping
+    @PostMapping //주문접수
     public ResponseEntity<?> orderReceived(@RequestBody OrderRequestDTO orderRequestDTO) {
         try {
             multiService.orderReceived(orderRequestDTO);
@@ -26,9 +29,16 @@ public class OrderController {
         }
     }
 
-    @GetMapping
+    @GetMapping //주문현황
     public ResponseEntity<?> checkOrder(){
         List<OrderResponseDTO> orderResponseDTOS = multiService.getOrder();
         return ResponseEntity.status(HttpStatus.OK).body(orderResponseDTOS);
     }
+
+    @MessageMapping("/api/orderList")
+    @SendTo("/api/sub/orderList")
+    public OrderResponseDTO realTimeOrder (OrderRequestDTO orderRequestDTO){
+        return multiService.getDto(Order.builder().name(orderRequestDTO.name()).count(orderRequestDTO.count()).build());
+    }
+
 }
