+++ frontend/src/app/Order/List/page.tsx
@@ -1,0 +1,18 @@
+'use client'
+
+import { getSocket } from "@/app/API/SocketAPI";
+import { useEffect, useState } from "react"
+import SockJS from "sockjs-client";
+
+function OrderList(){
+    const [socket,setSocket] =useState(null as any);
+
+    useEffect(() =>{
+        setSocket(getSocket([]));
+    },[])
+    return(
+        <div>
+
+        </div>
+    )
+}
